pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "qrem"
version = "0.1.5"
authors = [
  { name = "Jan Tuziemski, Filip Maciejewski, Joanna Majsak, Oskar Słowik, Marcin Kotowski, Katarzyna Kowalczyk-Murynka, Paweł Przewłocki, Piotr Podziemski, Michał Oszmaniec", email = "oszmaniec@cft.edu.pl" },
]
maintainers = [{ name = "Michał Oszmaniec", email = "oszmaniec@cft.edu.pl" }]
description = "QREM package provides a versatile set of tools for the characterization and mitigation of readout noise in NISQ quantum devices."
readme = "README.md"
keywords = [
  "quantum",
  "nisq",
  "quantum-devices",
  "quantum-physics",
  "physics",
  "sdk",
  "qiskit",
]
requires-python = ">=3.8"
classifiers = [
  "Development Status :: 3 - Alpha",
  "Environment :: Console",
  "Operating System :: Microsoft :: Windows",
  "Operating System :: MacOS",
  "Operating System :: POSIX :: Linux",
  "License :: OSI Approved :: Apache Software License",
  "Operating System :: OS Independent",
  "Intended Audience :: Developers",
  "Intended Audience :: Science/Research",
  "Topic :: Scientific/Engineering :: Physics",
  "Programming Language :: Python :: 3.7",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
]
dependencies = [
  "ipykernel >= 6.1.0",
  "python-dotenv >= 1.0.0",
  "orjson >= 3.9.10",
  "tqdm >= 4.64.0",
  "numpy >= 1.18.0, < 1.24",
  "scipy >= 1.7.0",
  "pandas >= 1.5.0",
  "picos >= 2.4.0",
  "qiskit >= 0.43.3",
  "qiskit-ibm-runtime >= 0.11.2",
  "qiskit-ibm-provider >= 0.6.1",
  "qiskit-braket-provider >= 0.0.3",
  "amazon-braket-sdk >= 1.61.0",
  "qutip >= 4.7.1",
  "matplotlib >= 3.6.0",
  "seaborn >= 0.12.0",
  "networkx >= 0.12.0, < 3.0",
  "sympy >= 1.11.0",
  "pyquil >= 4.0.3",
  "configargparse >= 1.5.0",
]

[project.optional-dependencies]
visualization = ["manim >= 0.17.2"]
test = ["pytest-cov ~=3.0.0"]

#[project.scripts]
#my-script = "my_package.module:function"

# [project.urls]
# "Homepage" = "https://github.com/cft-nisq/qrem"
# "Bug Tracker" = "https://github.com/cft-nisq/qrem/issues"
# "Documentation" = "https://github.com/cft-nisq/qrem/docs"
# "Tutorials" = "https://github.com/cft-nisq/qrem/tutorials"

[tool.setuptools.packages.find]
where = ["src"] # list of folders that contain the packages (["."] by default)
